pyproject.toml
[project]
name = "chatgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gradio>=4.44.1",
    "matplotlib>=3.9.4",
    "openai>=1.107.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "xlrd>=2.0.2",
]
